Add signIn and signOut functions
