+++ src/App.js
@@ -1,26 +1,22 @@
 import {db, googleAuthProvider, firebase} from './config/firebase';
 import Button from './componets/Button';
+import {signInWithGoogle, signOut} from './login';
+import sigInwithGoogle from './login/signInWithGoogle';
 
 
 function App() {
-  
-  const sigInwithGoogle = async () => {
-    //coloca el lenguaje de preferencia del dispositivo
-    firebase.auth().useDeviceLanguage();
-
-    //Inicio el procedo de login dentro de try ... catch
-    try{
-      await firebase.auth().signInWithPopup(googleAuthProvider);
-
-    } catch (e) {
-      console.error(e.message);
-        }
-  };
-
 // Renderizar el componente Button
   return (
     <div>
-      <button onClick={sigInwithGoogle}> Sing in with Google</button>
+      {
+        user ? (
+          <>
+          <Button onclick={signOut}> Sing in with Google</Button>
+          <p>Bienvenido al CHAT</p>
+          </>
+        ) :<Button onClick={sigInwithGoogle}> Sing in with Google</Button>
+      }
+
     </div>
   );
 }
