+++ src/login/index.js
@@ -1,0 +1,7 @@
+import sigInwithGoogle from './signInWithGoogle';
+import signOut from './signOut';
+
+export {
+    signInWithGoogle,
+    signOut,
+}
